make auth helper pluggable
